FIX: Task assembly file still in use causing removal failure
After running completes, the example contains a task that removes the
assembly of the task itself so Unity doesn't pick it up (because it
generates errors for it). This is problematic on Windows, because the
assembly DLL file is still in use (loaded) by the MSBuild process at
that point. To fix this, keep the DLL, but mark it as ignored by Unity.