FIX: Fix built-in assembly detection failing when Library/Bee is not generated
I don't understand when this folder is generated. I have tested on Linux
and Windows and it existed immediately after opening Unity once. Then I
tested another project on Windows recently, and now it isn't generated.

There seems to be no other way to detect the assemblies from the project
folder or the Unity base installation path reliably, so require the user
setting it once as next best solution.

This also refactors async LINQ code to be sequential instead of
parallellized as it was making console output not be sequential any more
due to these changes, and most of it is disk I/O which doesn't
parallellize well anyway.