FIX: UWP builds fail due to built-in assemblies being impossible to fully ignore
We don't want to drop the files as tools such as OmniSharp might still
want to scan the assemblies for IntelliSense, but Unity insists on still
using these in its reference rewriter on UWP builds, even when they
are excluded from all platforms, are not referenced automatically, are
set to not validate, and have define constraints.

The *only* way to ensure Unity doesn't do so is to fully delete them,
which in turn might pose problems for OmniSharp, but we have no better
option, currently.

Partially reverts 9a37fc03c8fcf29d4b4b10abe38c234614070c28.