FIX: Fix OmniSharp complaining about missing meta files for built-in packages
This is an attempt at an alternative solution to 9a37fc03c8fcf29d4b4b10abe38c234614070c28., which was reverted in a4b43da14b42dcb4e25d8260ac7aa34b70ea2a80.

We still drop the assemblies, as we can't keep (or exclude) them due to
this breaking UWP IL2CPP builds on Unity 2021.3, but we keep the rest of
these folders, as well as the ref folders of other packages, so the
matching XML database files stay around. OmniSharp started complaining
about these XML files being missing when analyzing them in some cases,
which this seems to fix.