FEAT: #3 Support generating Unity asmdefs per package to cherry-pick analyzers
In our Unity projects, we currently place a RoslynAnalyzerWrapper.asmdef
in the parent folder of the NuGet package folder to ensure that all
Roslyn analyzers inside the NuGet packages (which are already
automatically tagged as RoslynAnalyzer in Unity) don't automatically
apply to every package/asmdef in your Unity project (such as your own
scripts), which would otherwise also include Unity's own packages,
subsequently causing Unity editor errors with the Unity test framework
package for some of our Meziantou rules. Whilst the higher-level asmdef
works fine for our purposes, it prevents actual source generators such
as ZeroQL from being applied to other assembly definitions, despite
being tagged correctly as RoslynAnalyzer.

The 'fix' Unity prescribes is to reference the
RoslynAnalyzerWrapper.asmdef in other asmdefs you want the analyzers to
apply to - doing so will make them transient and also let them apply to
the referencing assembly. However, with the higher-level asmdef this
would then also apply all Roslyn analyzers such as Roslynator to our own
asmdefs again, which was part of the reason we were doing this (we don't
want Unity to report or block on Roslyn analyzer errors of our code
analysis rulesets as we already see them in our IDE, it slows down
Unity, and it is annoying during development).

By creating a single asmdef during generation, we alleviate the need to
manually include a higher-level asmdef, and users can cherry-pick which
analyzer and source generators from which NuGet packages they want and
don't want to apply to which of their own asmdefs.