FIX: Generate NuGet package asmdef meta files to provide stable GUIDs
The problem with letting Unity generate them is that every user will
get their own GUIDs for the asmdef meta files, as the NuGet package
folder is ideally not in Git. This makes Unity lose references to these
assemblies for project assemblies because the GUIDs keep changing.

Generate stable GUIDs for these files based on the package name (and
version if there are multiple) so that every user gets/restores the same
GUIDs and they also remain the same when changing NuGet package
versions.